# backend/docs/openapi.yaml
openapi: "3.1.0"
info:
  title: NyumbaLink API
  version: "2.0.0"
  description: |
    ## NyumbaLink — AI-Powered Accommodation Brokerage System
    **Dar es Salaam, Tanzania**

    NyumbaLink connects **Seekers** (tenants) with verified **Landlords** across
    Dar es Salaam's five municipalities — Kinondoni, Ilala, Ubungo, Temeke, and Kigamboni.

    ### Authentication
    Protected endpoints require a **Bearer JWT** in the `Authorization` header.
    ```
    Authorization: Bearer <access_token>
    ```
    Obtain tokens via `POST /api/v1/auth/login`.
    Access tokens expire after **24 hours**.
    Use `POST /api/v1/auth/refresh` to renew.

    ### Roles
    | Role | Description |
    |---|---|
    | `SEEKER` | Tenant searching for accommodation |
    | `LANDLORD` | Property owner managing listings |
    | `ADMIN` | Platform administrator |

    ### Base URL
    - Development: `http://localhost:8000`
    - Production: `https://api.nyumbalink.co.tz`
  contact:
    name: NyumbaLink Support
    email: mnemosadick@gmail.com
    url: https://nyumbalink.co.tz
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.nyumbalink.co.tz
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token. Obtain via `POST /api/v1/auth/login`.
        Include as: `Authorization: Bearer <token>`
  schemas:
...
    PaymentInitiate:
      type: object
      required: [property_id, amount, payment_method]
      properties:
        property_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        amount:
          type: number
          format: float
          minimum: 1
          example: 640000
        payment_method:
          type: string
          enum: [MPESA, AIRTEL, TIGO, HALOPESA, CARD]
          example: "MPESA"
        phone_number:
          type: string
          nullable: true
          example: "+255653870753"
          description: "Required for mobile money payments"
        description:
          type: string
          nullable: true
          example: "Rent payment — Kiota Hostel July 2025"
    PaymentOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        property_id:
          type: string
          format: uuid
          nullable: true
        amount:
          type: number
          format: float
          nullable: true
          example: 640000
        payment_method:
          type: string
          nullable: true
          example: "MPESA"
        transaction_reference:
          type: string
          nullable: true
          example: "TXNABC123456789"
        payment_status:
          type: string
          nullable: true
          enum: [PENDING, COMPLETED, FAILED, REFUNDED]
          example: "COMPLETED"
        paid_at:
          type: string
          format: date-time
          nullable: true
    ReceiptOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        payment_id:
          type: string
          format: uuid
        receipt_number:
          type: string
          nullable: true
          example: "RCT87654321"
        receipt_url:
          type: string
          nullable: true
          example: "https://receipts.nyumbalink.co.tz/550e8400-e29b-41d4-a716-446655440002"
        generated_at:
          type: string
          format: date-time

    # ── Reviews ───────────────────────────────────────────────────────────────
...
    ConversationCreate:
      type: object
      required: [landlord_id]
      properties:
        landlord_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440003"
    MessageOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        sender_id:
          type: string
          format: uuid
        message:
          type: string
          nullable: true
          example: "Is the room still available?"
        message_type:
          type: string
          enum: [TEXT, IMAGE, FILE]
          example: "TEXT"
        is_read:
          type: boolean
          example: false
        sent_at:
          type: string
          format: date-time
    ConversationOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        seeker_id:
          type: string
          format: uuid
        landlord_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/MessageOut"
    MessageCreate:
      type: object
      required: [conversation_id, message]
      properties:
        conversation_id:
          type: string
          format: uuid
        message:
          type: string
          minLength: 1
          example: "What is the deposit amount?"
        message_type:
          type: string
          enum: [TEXT, IMAGE, FILE]
          default: "TEXT"

    # ── Notifications ─────────────────────────────────────────────────────────
...
    PricePredictionRequest:
      type: object
      required: [ward, water_reliability, electricity_config, is_furnished, has_wifi, distance_to_transit_m]
      properties:
        ward:
          type: string
          enum: [Mikocheni, Mwananyamala, Kijitonyama, Mwenge, Mabatini, Makumbusho, Sinza]
          example: "Mikocheni"
        structural_rooms:
          type: integer
          minimum: 1
          default: 1
          example: 1
        water_reliability:
          type: string
          enum: [none, intermittent, daily, continuous]
          example: "daily"
        electricity_config:
          type: string
          enum: [none, luku, fixed, solar]
          example: "luku"
        distance_to_transit_m:
          type: number
          format: float
          minimum: 0
          example: 400
        floor_area_sqm:
          type: number
          format: float
          minimum: 5
          default: 25
          example: 18
        is_furnished:
          type: boolean
          default: false
        has_wifi:
          type: boolean
          default: false
    PricePredictionResponse:
      type: object
      properties:
        estimated_fair_market_value:
          type: number
          format: float
          example: 165000
        calculated_variance_range:
          type: object
          properties:
            minimum_fair_boundary:
              type: number
              format: float
              example: 145200
            maximum_fair_boundary:
              type: number
              format: float
              example: 184800
        confidence_score_metric:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.91
        ward:
          type: string
          example: "Mikocheni"
        is_overpriced:
          type: boolean
          example: false
        overpriced_by_percent:
          type: number
          format: float
          example: 0.0

    # ── NLP Chatbot ───────────────────────────────────────────────────────────
...
paths:
...
  /api/v1/payments:
    post:
      tags: [Payments]
      summary: Initiate a payment
      description: |
        Initiate payment via mobile money or card.
        Supported methods: **MPESA**, **AIRTEL**, **TIGO**, **HALOPESA**, **CARD**.
        Auto-generates a receipt on completion.
        In production, wire to Vodacom / Airtel push-pay APIs.
      operationId: payments_initiate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInitiate"
            examples:
              mpesa:
                summary: M-Pesa payment
                value:
                  property_id: "550e8400-e29b-41d4-a716-446655440001"
                  amount: 640000
                  payment_method: "MPESA"
                  phone_number: "+255653870753"
                  description: "July 2025 rent — Kiota Hostel"
      responses:
        "201":
          description: Payment initiated and completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOut"
...
  /api/v1/payments/{payment_id}/receipt:
    get:
      tags: [Payments]
      summary: Get payment receipt
      operationId: payments_get_receipt
      parameters:
        - name: payment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptOut"
        "404":
          description: Receipt not found

  # ────────────────────────────────── REVIEWS ────────────────────────────────
...
  /api/v1/messaging/conversations:
    post:
      tags: [Messaging]
      summary: Start or retrieve a conversation with a landlord
      description: |
        Creates a new conversation between the authenticated seeker and the specified landlord.
        If a conversation already exists, returns the existing one (idempotent).
      operationId: messaging_start_conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationCreate"
      responses:
        "201":
          description: Conversation created or retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationOut"
    get:
      tags: [Messaging]
      summary: List all my conversations
      operationId: messaging_list_conversations
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConversationOut"
  /api/v1/messaging/conversations/{conversation_id}:
    get:
      tags: [Messaging]
      summary: Get a conversation with all its messages
      operationId: messaging_get_conversation
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationOut"
        "403":
          description: Not your conversation
        "404":
          description: Not found
  /api/v1/messaging/messages:
    post:
      tags: [Messaging]
      summary: Send a message in a conversation
      operationId: messaging_send_message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageCreate"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageOut"
  /api/v1/messaging/messages/{message_id}/read:
    patch:
      tags: [Messaging]
      summary: Mark a message as read
      operationId: messaging_mark_read
      parameters:
        - name: message_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Marked as read

  # ──────────────────────────── NOTIFICATIONS ────────────────────────────────
...
  /api/v1/predict-price/:
    post:
      tags: [Price Intelligence]
      summary: Predict fair market rental value
      description: |
        XGBoost hedonic pricing model. Predicts fair market rent based on:
        ward, rooms, utilities, distance to transit, and furnishing.
        If the listed price is higher than predicted by >10%, `is_overpriced` is set.
      security: []
      operationId: price_predict
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PricePredictionRequest"
            example:
              ward: "Mikocheni"
              structural_rooms: 1
              water_reliability: "daily"
              electricity_config: "luku"
              distance_to_transit_m: 300
              floor_area_sqm: 18
              is_furnished: false
              has_wifi: false
      responses:
        "200":
          description: Price prediction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricePredictionResponse"
  /api/v1/predict-price/flag-listing:
    post:
      tags: [Price Intelligence]
      summary: Compare listed price against ML estimate
      description: Returns `is_overpriced=true` if listed price exceeds prediction by >10%.
      security: []
      operationId: price_flag_listing
      parameters:
        - name: listed_price
          in: query
          required: true
          schema:
            type: number
            format: float
          example: 300000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PricePredictionRequest"
      responses:
        "200":
          description: Price flag result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricePredictionResponse"

  # ────────────────────────────── CONVERSATION ───────────────────────────────
...